Fixed the issue with keywords
Fixed the issue with keywords, i.e., there id was not getting updated in the keywords_to_articles table. Also solved the problem in which the older tags/keywords were re-written in the respective table. However in doing so, I may have hit upon a snag. When I give new list of tags and keywords the application fails. The server response says "A database constrain has failed". The error doesn't show up in the subsequent submits with the same tags and keywords. 

Solve the issue.